requirements.txt
# PDF Processing
PyPDF2==3.0.1

# LangChain and Vector Store
langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.2

# Astra DB
astrapy==0.7.0

# Embeddings (OpenAI)
openai==1.7.0

# Environment Variables
python-dotenv==1.0.0

# Text Processing
tiktoken==0.5.2



